backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
axios==0.4.0
brew==0.1.3
build==1.2.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
Faker==26.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
openai==1.35.13
orjson==3.10.6
packaging==24.1
parsimonious==0.10.0
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PyMySQL==1.1.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
replicate==0.32.0
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
shellingham==1.5.4
singlestoredb==1.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlalchemy_singlestoredb==1.1.0
sqlparams==6.0.1
starlette==0.37.2
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
tqdm==4.66.4
transformers==4.44.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zipp==3.19.2
